+++ CHANGELOG.md
@@ -10,10 +10,12 @@
 - [FP-1297 - Stay in app on logout](https://movai.atlassian.net/browse/FP-1297)
 - [FP-1509 - Saving a new document will now reload the doc](https://movai.atlassian.net/browse/FP-1509)
 - [FP-1538 - Color between tabs now, more clearer](https://movai.atlassian.net/browse/FP-1538)
+- [FP-1548 - Added testids to help QA](https://movai.atlassian.net/browse/FP-1548)
 - [FP-1560 - Saving an untitled file will now always prompt for a name](https://movai.atlassian.net/browse/FP-1560)
 - [FP-1595 - Removed unused function that contained vulnerabilities](https://movai.atlassian.net/browse/FP-1595)
 - [FP-1610 - Remove editor keybinds when editor is blurred](https://movai.atlassian.net/browse/FP-1610)
-- [FP-1548 - Added testids to help QA](https://movai.atlassian.net/browse/FP-1548)
+- [FP-1838 - Give focus to active editor if it becomes dirty](https://movai.atlassian.net/browse/FP-1838)
+- [FP-1978 - Add Description field when Adding/Editing parameter on Flow Template](https://movai.atlassian.net/browse/FP-1978)
 
 **Callback Editor**